+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/CreateTest.cs
@@ -19,7 +19,7 @@ public class CreateTest : BaseMockServerTest
                 "type": "contentStrategy",
                 "attributes": {
                   "name": "Featured Travel",
-                  "filter": "HIGHEST_CASHBACK",
+                  "sort": "HIGHEST_CASHBACK",
                   "categories": [
                     "Travel"
                   ],
@@ -41,7 +41,7 @@ public class CreateTest : BaseMockServerTest
               "attributes": {
                 "name": "Featured Travel",
                 "organizationId": "org-123",
-                "filter": "HIGHEST_CASHBACK",
+                "sort": "HIGHEST_CASHBACK",
                 "categories": [
                   "Travel"
                 ],
@@ -80,7 +80,7 @@ public class CreateTest : BaseMockServerTest
                     Attributes = new CreateContentStrategyAttributes
                     {
                         Name = "Featured Travel",
-                        Filter = ContentStrategyFilter.HighestCashback,
+                        Sort = ContentStrategySort.HighestCashback,
                         Categories = new List<CategoryOption>() { CategoryOption.Travel },
                         CategoryExclusions = new List<CategoryOption>() { CategoryOption.Gas },
                         MerchantExclusions = new List<string>() { "merchant-abc" },