+++ src/KardFinancial.Test/Unit/MockServer/Organizations/ContentStrategies/UpdateTest.cs
@@ -43,7 +43,7 @@ public class UpdateTest : BaseMockServerTest
               "attributes": {
                 "name": "name",
                 "organizationId": "organizationId",
-                "filter": "NEWLY_LIVE",
+                "sort": "NEWLY_LIVE",
                 "categories": [
                   "Arts & Entertainment",
                   "Arts & Entertainment"