+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/GetTest.cs
@@ -31,7 +31,7 @@ public class GetTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filter": "NEWLY_LIVE",
+                    "sort": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"
@@ -52,7 +52,7 @@ public class GetTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filter": "NEWLY_LIVE",
+                    "sort": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"