+++ src/KardFinancial.Test/Unit/MockServer/Organizations/Placements/ListTest.cs
@@ -43,7 +43,7 @@ public class ListTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filter": "NEWLY_LIVE",
+                    "sort": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"
@@ -64,7 +64,7 @@ public class ListTest : BaseMockServerTest
                   "attributes": {
                     "name": "name",
                     "organizationId": "organizationId",
-                    "filter": "NEWLY_LIVE",
+                    "sort": "NEWLY_LIVE",
                     "categories": [
                       "Arts & Entertainment",
                       "Arts & Entertainment"