+++ src/KardFinancial.Test/Unit/Serialization/CreateContentStrategyAttributesTest.cs
@@ -16,7 +16,7 @@ public class CreateContentStrategyAttributesTest
         var json = """
             {
               "name": "Featured Travel",
-              "filter": "HIGHEST_CASHBACK",
+              "sort": "HIGHEST_CASHBACK",
               "categories": [
                 "Travel"
               ],
@@ -31,7 +31,7 @@ public class CreateContentStrategyAttributesTest
         var expectedObject = new CreateContentStrategyAttributes
         {
             Name = "Featured Travel",
-            Filter = ContentStrategyFilter.HighestCashback,
+            Sort = ContentStrategySort.HighestCashback,
             Categories = new List<CategoryOption>() { CategoryOption.Travel },
             CategoryExclusions = new List<CategoryOption>() { CategoryOption.Gas },
             MerchantExclusions = new List<string>() { "merchant-abc" },
@@ -46,7 +46,7 @@ public class CreateContentStrategyAttributesTest
         var inputJson = """
             {
               "name": "Featured Travel",
-              "filter": "HIGHEST_CASHBACK",
+              "sort": "HIGHEST_CASHBACK",
               "categories": [
                 "Travel"
               ],